requirements.txt
click==7.1.2
Deprecated==1.2.13
Flask==1.1.4
gevent==22.10.2
greenlet==2.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
opentelemetry-api==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
typing-extensions==4.4.0
Werkzeug==1.0.1
wrapt==1.14.1
zope.event==4.6
zope.interface==5.5.2
